version3

+++ eval1/script.sh
@@ -1,22 +1,31 @@
 #!/bin/bash
-touch /tmp/out/lock
+file="/tmp/in/"
 
-DIR="/etc/tmp/out/"
+DIR="/etc/tmp/out"
 
-file="/tmp/out/lock"
+lock="/tmp/in/lock"
 
-if [ -d "$DIR" ]
+if [ -d "!$DIR" ]
 then
-	if [ -f "$file" ]
-	then
-        	gzip -r /tmp/in/*
-        	mv /tmp/in/* /tmp/out/
-        	rm /tmp/out/lock
-	else
-        	exit 22
-	fi
+	mkdir /tmp/out
 else
-	echo "Ce dossier n'existe pas"
-	exit 3
+	touch /tmp/in/lock
 fi
 
+if [ ! -f "$lock" ]
+then
+	for files in /tmp/in/* 
+	do
+	if [[ -f $file && -r $file ]]
+	then
+        	gzip -r /tmp/in/*
+        	mv $files.gz /tmp/out/
+        	rm /tmp/out/lock
+	else
+		echo "File is unable"
+	fi
+	done
+else
+	echo "Script already in use"
+        exit 22
+fi
